Move util functions to their own file
